+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the `rename` command which renames contracts
+- Added the `pack` command which builds and prepares a publish-ready package of contracts' wrappers
 
 ## [0.33.1] - 2025-05-16